feat(plugin): gate dynamic loading with plugins feature
