+++ Cargo.toml
@@ -26,7 +26,7 @@ log = "0.4"
 env_logger = "0.10"
 
 # Optional dependencies
-libloading = "0.8"
+libloading = { version = "0.8", optional = true }
 
 [dev-dependencies]
 tempfile = "3.8"
