--- src/plugin.rs
@@ -1,202 +1,0 @@
-use anyhow::{Result, Context};
-use colored::*;
-use indicatif::{ProgressBar, ProgressStyle, MultiProgress};
-use pager::Pager;
-use prettytable::{Table, row, format};
-use std::{path::PathBuf, collections::HashMap};
-use crate::{paging::MemoryImage, loader::load_memory_image};
-
-/// Represents a finding from a memory forensics plugin
-#[derive(Debug, Clone)]
-pub struct Finding {
-    pub plugin: String,
-    pub addr: u64,
-    pub desc: String,
-    pub confidence: u8, // 0-100 confidence level
-}
-
-/// Core trait for memory forensics plugins
-pub trait MemoryPlugin {
-    fn name(&self) -> &'static str;
-    fn description(&self) -> &'static str;
-    fn scan(&self, img: &MemoryImage, progress: &ProgressBar) -> Vec<Finding>;
-}
-
-// A simple string carving plugin that looks for ASCII and UTF-16 strings in memory
-pub struct StringCarvePlugin;
-
-impl MemoryPlugin for StringCarvePlugin {
-    fn name(&self) -> &'static str {
-        "string_carve"
-    }
-    
-    fn description(&self) -> &'static str {
-        "Scans memory for ASCII and UTF-16 strings"
-    }
-
-    fn scan(&self, img: &MemoryImage, progress: &ProgressBar) -> Vec<Finding> {
-        let mut findings = Vec::new();
-        let size = img.size();
-        
-        // For demonstration purposes, we'll simulate finding strings
-        // In a real scanner, we'd look for MIN_STRING_LEN consecutive printable chars
-        let min_string_len = 8;
-        // In a real implementation, we would scan through the memory image
-        // looking for sequences of printable characters
-        
-        // Set up progress bar
-        progress.set_length(size as u64);
-        progress.set_message("Scanning for strings");
-        
-        let scan_points = [
-            (0x1000, "Password: admin123"),
-            (0x2500, "config.xml"),
-            (0x5000, "http://example.com/data"),
-            (0x7A00, "SELECT * FROM users"),
-            (0xA000, "/etc/shadow"),
-            (0xC000, "ssh-rsa AAAA..."),
-            (0xE000, "SECRET_KEY=abc123"),
-            (0xF500, "192.168.1.1"),
-        ];
-        
-        // Iterate through chunks of memory (simulated here)
-        for chunk_start in (0..size).step_by(4096) {
-            // Update progress every 4KB
-            progress.set_position(chunk_start as u64);
-            
-            // Check if any of our simulated strings are in this chunk
-            for &(addr, string) in scan_points.iter() {
-                if addr >= chunk_start && addr < chunk_start + 4096 {
-                    // Only include strings that meet our minimum length requirement
-                    if string.len() >= min_string_len {
-                        findings.push(Finding {
-                            plugin: self.name().to_string(),
-                            addr: addr as u64,
-                            desc: string.to_string(),
-                            confidence: 90,
-                        });
-                    } else {
-                        // For shorter strings, lower confidence
-                        findings.push(Finding {
-                            plugin: self.name().to_string(),
-                            addr: addr as u64,
-                            desc: string.to_string(),
-                            confidence: 50, // Lower confidence for short strings
-                        });
-                    }
-                }
-            }
-            
-            // In a real scanner, we'd do something like this:
-            // if let Some(chunk_bytes) = img.get_bytes(chunk_start, 4096) {
-            //     scan_for_strings(chunk_bytes, chunk_start, min_string_len, &mut findings);
-            // }
-            
-            // Simulate work
-            if chunk_start % (1024 * 1024) == 0 {  // Every 1MB
-                std::thread::sleep(std::time::Duration::from_millis(10));
-            }
-        }
-        
-        progress.finish_with_message(format!("Found {} strings", findings.len()));
-        findings
-    }
-}
-
-// Registry of available plugins
-struct PluginRegistry {
-    plugins: HashMap<String, Box<dyn MemoryPlugin>>,
-}
-
-impl PluginRegistry {
-    fn new() -> Self {
-        let mut registry = Self {
-            plugins: HashMap::new(),
-        };
-        
-        // Register built-in plugins
-        registry.register(Box::new(StringCarvePlugin));
-        
-        registry
-    }
-    
-    fn register(&mut self, plugin: Box<dyn MemoryPlugin>) {
-        self.plugins.insert(plugin.name().to_string(), plugin);
-    }
-    
-    fn get(&self, name: &str) -> Option<&Box<dyn MemoryPlugin>> {
-        self.plugins.get(name)
-    }
-    
-    fn list_plugins(&self) -> Vec<&str> {
-        self.plugins.keys().map(|k| k.as_str()).collect()
-    }
-}
-
-pub fn run_plugin(dump_path: PathBuf, plugin_name: String) -> Result<()> {
-    println!("{} {} {} {}",
-        "Running plugin".bright_green(),
-        plugin_name.bright_yellow().bold(),
-        "on".bright_green(),
-        dump_path.display().to_string().bright_cyan()
-    );
-    
-    // Create plugin registry
-    let registry = PluginRegistry::new();
-    
-    // Check if plugin exists
-    let plugin = registry.get(&plugin_name)
-        .with_context(|| format!("Plugin '{}' not found. Available plugins: {}", 
-            plugin_name, 
-            registry.list_plugins().join(", ")
-        ))?;
-    
-    println!("{}: {}", "Plugin description".bright_blue(), plugin.description());
-    
-    // Load memory image
-    let memory_image = load_memory_image(&dump_path)?;
-    
-    // Set up progress bars
-    let multi_progress = MultiProgress::new();
-    let scan_progress = multi_progress.add(ProgressBar::new(100));
-    scan_progress.set_style(ProgressStyle::with_template(
-        "{spinner:.green} [{elapsed_precise}] [{bar:40.cyan/blue}] {bytes}/{total_bytes} {msg}"
-    )?.progress_chars("#>-"));
-    
-    // Run the plugin
-    println!("{}", "Starting scan...".bright_green());
-    let findings = plugin.scan(&memory_image, &scan_progress);
-    
-    // Display findings using pager if there are many
-    if !findings.is_empty() {
-        // Create a table for the findings
-        let mut table = Table::new();
-        table.set_format(*format::consts::FORMAT_NO_BORDER_LINE_SEPARATOR);
-        table.set_titles(row![b->"Address", b->"Confidence", b->"Description"]);
-        
-        for finding in &findings {
-            table.add_row(row![
-                format!("0x{:08X}", finding.addr),
-                format!("{}%", finding.confidence),
-                finding.desc
-            ]);
-        }
-        
-        // Use pager if we have many findings
-        if findings.len() > 20 {
-            Pager::new().setup();
-        }
-        
-        println!("\n{} {} {}", 
-            "Found".bright_green(),
-            findings.len().to_string().bright_yellow().bold(),
-            "items".bright_green()
-        );
-        
-        table.printstd();
-    } else {
-        println!("{}", "No findings from the scan".bright_yellow());
-    }
-    
-    Ok(())
-}
